// apps/frontend-web/app/api/generated/model/fetchKitchensResponseDtoKitchensItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * AoPonto FoodService API
 * Documentação da API do sistema AoPonto
 * OpenAPI spec version: 1.0
 */

export type FetchKitchensResponseDtoKitchensItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string
  description: string
  ip: string
  port: string
  printer: string
  versionReg?: number | null
  versionSync?: number | null
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string
  updatedAt?: string | null
}